refactor: seperate find one summary function
